implement new multi file reader interface
