+++ CMakeLists.txt
@@ -38,7 +38,12 @@ set(LOADABLE_EXTENSION_NAME ${TARGET_NAME}_loadable_extension)
 project(${TARGET_NAME})
 include_directories(src/include)
 
-set(EXTENSION_SOURCES src/avro_extension.cpp)
+set(
+	EXTENSION_SOURCES
+	src/avro_extension.cpp
+	src/avro_reader.cpp
+	src/avro_multi_file_info.cpp
+)
 
 build_static_extension(${TARGET_NAME} ${EXTENSION_SOURCES})
 build_loadable_extension(${TARGET_NAME} " " ${EXTENSION_SOURCES})
