+++ test/sql/avro.test
@@ -102,7 +102,7 @@ favorite_color	VARCHAR
 statement error
 from read_avro(['test/userdata1.avro', 'test/users.avro'])
 ----
-Schema of file
+schema mismatch in glob
 
 # example from readme
 query III
@@ -334,7 +334,7 @@ avro_schema	UNION(u0 STRUCT("data" BLOB), u1 STRUCT(fatal BOOLEAN, "name" VARCHA
 query I
 FROM read_avro('test/query_small.avro')
 ----
-{'data': 100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A100,200,300,400\x0A300,400,500,600\x0A}
+{'data': '100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A100,200,300,400\\x0A300,400,500,600\\x0A'}
 {'bytesScanned': 1024, 'totalBytes': 1024}
 {'totalBytes': 1024}
 
@@ -415,14 +415,20 @@ comments	VARCHAR
 statement error
 FROM read_avro('test/union-name-*.avro', filename=true, union_by_name=true) order by one;
 ----
-Not implemented Error: union_by_name for Avro reads
+Not implemented Error: 'union_by_name' not implemented for Avro reader yet
 
 
+# FIXME: "need to"?
 # need to fail on differing types
-statement error
+query IIII
 FROM read_avro('test/union-name-*.avro', filename=true) order by one;
 ----
-Schema of file
+10	2.0	s30	test/union-name-1.avro
+11	2.1	s31	test/union-name-1.avro
+12	2.2	s32	test/union-name-2.avro
+13	2.3	s33	test/union-name-2.avro
+14	2.4	s34	test/union-name-3.avro
+15	2.5	s35	test/union-name-3.avro
